+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Highlights are marked with a pancake 🥞
 ### Changed
 
 - One global state object and operation graph for all groups [#781](https://github.com/p2panda/p2panda/pull/781)
+- Handle strong removal edge cases and cycles in nested groups, improve concurrent re-adds [#788](https://github.com/p2panda/p2panda/pull/788)
 
 ### Fixed